Use cmake to build mzgl
Fix version and includes

Fixes for using with opt

Dont link to font stash

Make sure deps are public

make it build with cmake!!

Fix some naming

Clean up deps

Update platform types

ios compilation

Update cmake properties

Use our fork of fontstash

Fix missing ios includes

roll back standard so that pytorch compiles

Send only to a single midi device

Use NEON horizontal max rather than std::max

Fix android

fix access of temporary memory in midi message output

Fix a crash when size is <= 1

Less verbose version

trash

variable

get and set value on variable

add program change in midi message

added inc/dec key codes

variable now works properly

Update lib/mzgl/util/Variable.h


Update lib/mzgl/util/Variable.h


updated as per PR

listenable now has scoped listener

Update cmake

added copy constructor and copy assigment to Variable

moved tests to mzgl, including variableTest - second test fails

Fix variable watcher

added getter setter and fix to constructor of variablewatcher

Ask speex to skip zeros

moved AllMidiDevices Impls to own files

Add helpers for the tests

improve listenable and add tests, and use it with AllMidiDevices

pr fixes

changed method signature in definition to match declaration

public enable_shared_from_this

added missing virtual function to rtmidi impl

setThreadName stub for windows

slight tidy of util

missing atomic include

androidFileDialog stubs

finished android file dialog

refactor util

fix test linker error

fixed file save dialog I think

found double instantiation of viewcontroller whoops

found why koala wasn't destructing

winfix2

missing include for std::size_t

Send and receive midi on android

Pass through the device

Input is working

Make it build with inputs and outputs

using shared_ptr

SEt main thread runenr

Use virtual overrides

const

Fix audio extracted from video crashing

Add invalid file tests

Make sure all memory is released

Refactor so all loading is shared

Test permissions and existance

Fix the file path representation and add more logging

Fix some tests

Fix windows test

Update windows tests

Fix missing include

fixed bug where midiReceived was searching through outputs rather than inputs

include optional for windows

Use cmake to build mzgl

Fix version and includes

Fixes for using with opt

Dont link to font stash

Make sure deps are public

make it build with cmake!!

Fix some naming

Clean up deps

Update platform types

ios compilation

Update cmake properties

Use our fork of fontstash

Fix missing ios includes

roll back standard so that pytorch compiles

Send only to a single midi device

Use NEON horizontal max rather than std::max

Fix android

fix access of temporary memory in midi message output

Fix a crash when size is <= 1

Less verbose version

trash

variable

get and set value on variable

add program change in midi message

added inc/dec key codes

variable now works properly

Update lib/mzgl/util/Variable.h


Update lib/mzgl/util/Variable.h


updated as per PR

listenable now has scoped listener

Update cmake

Fix midi includes

It builds and runs!

Not sure if this works

print in pretty colours

Update colouring of text

Fix the build

Add missing includes

Update lib/mzgl/LiveCodeApp.h

Update lib/mzgl/audio/AudioFileAndroid.cpp

Update lib/mzgl/gl/Font.cpp

Update lib/mzgl/util/winUtil.cpp

Fix source and resampler

Use raw root files

Dont all build AudioShare

Remove choc

Remove file system

Remove glfw

Remove glm

Remove open ssl

remove port audio

Remove rtmidi

Clean up the lib dir

Clean up the printing

FIx include paths

Add plugin lib

Make Android compile

Clean up the cmake

Fix midi compilation

+++ lib/mzgl/test/midiParserTests.cpp
@@ -1,5 +1,5 @@
 #include "tests.h"
-#include "MidiMessageParser.h"
+#include <mzgl/midi/MidiMessageParser.h>
 
 SCENARIO("Input sysex is parsed properly", "[midi-parser]") {
 	GIVEN("Some sysex and a parser") {
@@ -14,7 +14,7 @@ SCENARIO("Input sysex is parsed properly", "[midi-parser]") {
 		}};
 
 		WHEN("The data is parsed") {
-			parser.parse(sysex, 666, 0, 0);
+			parser.parse(sysex, static_cast<uint64_t>(666), static_cast<int32_t>(0), static_cast<int32_t>(0));
 			THEN("We should have got the callback") {
 				REQUIRE(gotCallback);
 			}
@@ -38,8 +38,8 @@ SCENARIO("Input sysex is parsed properly in parts", "[midi-parser]") {
 		}};
 
 		WHEN("The data is parsed in parts") {
-			parser.parse(sysex1, 666, 0, 0);
-			parser.parse(sysex2, 666, 0, 0);
+			parser.parse(sysex1, static_cast<uint64_t>(666), static_cast<int32_t>(0), static_cast<int32_t>(0));
+			parser.parse(sysex2, static_cast<uint64_t>(666), static_cast<int32_t>(0), static_cast<int32_t>(0));
 			THEN("We should have got the callback") {
 				REQUIRE(gotCallback);
 			}
@@ -87,7 +87,7 @@ SCENARIO("Delimits multiple note ons off properly", "[midi-parser]") {
 
 		WHEN("The messages are sent") {
 			for (auto &msg: msgs) {
-				parser.parse(msg, 666, 0, 0);
+				parser.parse(msg, static_cast<uint64_t>(666), static_cast<int32_t>(0), static_cast<int32_t>(0));
 			}
 			THEN("We should have got the callback") {
 				REQUIRE(messageCounter == originalMessages.size());
